feat: Refactor app structure by moving main application logic to a new file
