+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import App from "@/App";
+import App from "./test";
 
 ReactDOM.createRoot(document.querySelector("#root")!).render(
   <React.StrictMode>
